chore
[skip ci]

+++ app/src/main/java/com/github/zly2006/zhihu/ui/components/AnswerHorizontalPager.kt
@@ -74,8 +74,7 @@ fun AnswerHorizontalOverscroll(
 
     fun dampedOffset(rawDelta: Float): Float {
         val sign = if (rawDelta >= 0) 1f else -1f
-        return sign * maxOverscrollPx *
-            tanh(abs(rawDelta) / (maxOverscrollPx * HORIZONTAL_DAMPING_FACTOR)).toFloat()
+        return sign * maxOverscrollPx * tanh(abs(rawDelta) / (maxOverscrollPx * HORIZONTAL_DAMPING_FACTOR))
     }
 
     Box(
